♻️ refactor(model-runtime): rename provider xxxModelId modules to modelId
The provider directory already scopes the module, so the claude/google/
kimi/openai/qwen/glm prefixes were redundant. Exported symbol names are
unchanged.
