+++ packages/model-runtime/src/core/contextBuilders/openai.ts
@@ -3,7 +3,7 @@ import { Buffer } from 'buffer.js';
 import type OpenAI from 'openai';
 import { toFile } from 'openai';
 
-import { disableStreamModels, systemToUserModels } from '../../providers/openai/openaiModelId';
+import { disableStreamModels, systemToUserModels } from '../../providers/openai/modelId';
 import type { ChatStreamPayload, OpenAIChatMessage, UserMessageContentPart } from '../../types';
 import { isDeepSeekThinkingEligibleModel } from '../../utils/modelParse';
 import { parseDataUri } from '../../utils/uriParser';
