+++ packages/model-runtime/src/index.ts
@@ -36,7 +36,7 @@ export * from './helpers';
 export { LobeAkashChatAI } from './providers/akashchat';
 export { LobeAntGroupAI } from './providers/antgroup';
 export { LobeAnthropicAI } from './providers/anthropic';
-export * from './providers/anthropic/claudeModelId';
+export * from './providers/anthropic/modelId';
 export { LobeAzureAI } from './providers/azureai';
 export { LobeAzureOpenAI } from './providers/azureOpenai';
 export { LobeBailianCodingPlanAI } from './providers/bailianCodingPlan';
@@ -48,7 +48,7 @@ export { LobeComfyUI } from './providers/comfyui';
 export { LobeDeepSeekAI } from './providers/deepseek';
 export { LobeGLMCodingPlanAI } from './providers/glmCodingPlan';
 export { LobeGoogleAI } from './providers/google';
-export * from './providers/google/googleModelId';
+export * from './providers/google/modelId';
 export { LobeGroq } from './providers/groq';
 export { LobeKimiCodingPlanAI } from './providers/kimiCodingPlan';
 export { LobeHubAI } from './providers/lobehub';
@@ -57,13 +57,13 @@ export { LobeMinimaxAI } from './providers/minimax';
 export { LobeMinimaxCodingPlanAI } from './providers/minimaxCodingPlan';
 export { LobeMistralAI } from './providers/mistral';
 export { LobeMoonshotAI } from './providers/moonshot';
-export { isKimiAlwaysPreserveThinkingModel } from './providers/moonshot/kimiModelId';
+export { isKimiAlwaysPreserveThinkingModel } from './providers/moonshot/modelId';
 export { LobeNebiusAI } from './providers/nebius';
 export { LobeNewAPIAI } from './providers/newapi';
 export { LobeOllamaAI } from './providers/ollama';
 export { LobeOllamaCloudAI } from './providers/ollamacloud';
 export { LobeOpenAI } from './providers/openai';
-export * from './providers/openai/openaiModelId';
+export * from './providers/openai/modelId';
 export { LobeOpenRouterAI } from './providers/openrouter';
 export { LobePerplexityAI } from './providers/perplexity';
 export { LobeQwenAI } from './providers/qwen';
