+++ packages/model-runtime/src/providers/moonshot/index.ts
@@ -19,7 +19,7 @@ import {
   isKimiPreserveThinkingModel,
   isKimiReasoningEffortModel,
   isKimiThinkingToggleModel,
-} from './kimiModelId';
+} from './modelId';
 
 export interface MoonshotModelCard {
   context_length?: number;
