+++ packages/model-runtime/src/providers/opencodeZen/index.ts
@@ -3,7 +3,7 @@ import { LOBE_DEFAULT_MODEL_LIST, ModelProvider } from 'model-bank';
 import { createRouterRuntime } from '../../core/RouterRuntime';
 import type { CreateRouterRuntimeOptions } from '../../core/RouterRuntime/createRuntime';
 import { detectModelProvider, processMultiProviderModelList } from '../../utils/modelParse';
-import { responsesAPIModels } from '../openai/openaiModelId';
+import { responsesAPIModels } from '../openai/modelId';
 import { resolveProviderRouteModels } from '../utils/resolveProviderRouteModels';
 
 const ZEN_BASE_URL = 'https://opencode.ai/zen/v1';
